+++ pkg/apihelper/doc.go
@@ -12,4 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package apihelper // import "yunion.io/x/onecloud/pkg/vpcagent/apihelper"
+package apihelper // import "yunion.io/x/onecloud/pkg/apihelper"
